+++ src/Model/Entity/Sport.php
@@ -5,6 +5,14 @@ namespace App\Model\Entity;
 
 use Cake\ORM\Entity;
 
+/**
+ * Sport Entity
+ *
+ * @property int $id
+ * @property string $sport_name
+ * @property \Cake\I18n\DateTime|null $created_at
+ * @property \Cake\I18n\DateTime|null $updated_at
+ */
 class Sport extends Entity
 {
     // Add custom methods or virtual fields if needed