+++ src/Model/Table/SportsTable.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 namespace App\Model\Table;
 
 use Cake\ORM\Table;
+use Cake\Validation\Validator;
 
 class SportsTable extends Table
 {
@@ -23,5 +24,27 @@ class SportsTable extends Table
             'created' => 'created_at',
             'modified' => 'updated_at',
         ]);
+    }
+
+    /**
+     * Default validation rules.
+     *
+     * @param \Cake\Validation\Validator $validator Validator instance.
+     * @return \Cake\Validation\Validator
+     */
+    public function validationDefault(Validator $validator): Validator
+    {
+        $validator
+            ->integer('id')
+            ->allowEmptyString('id', null, 'create');
+
+        $validator
+            ->scalar('sport_name')
+            ->maxLength('sport_name', 162)
+            ->requirePresence('sport_name', 'create')
+            ->notEmptyString('sport_name')
+            ->add('sport_name', 'unique', ['rule' => 'validateUnique', 'provider' => 'table']);
+
+        return $validator;
     }
 }